set: Added some methods
Added some convinience methods.

Join is a joining other set and return a new set that contains only elements present in both sets.

Union is a merging with other set.

LeftOuter is a joining other set and return a new set that contains only elements present in the left set.

RightOuter is a joining other set and retrun a new set that contains only elements present in the right set.